+++ tests/lex/regex.js
@@ -1,40 +1,40 @@
 define(['bennu/parse', 'khepri/lex/regular_expression_lexer'], function(parse, regular_expression_lexer){
     
     var reEq = function(x, y) {
-        return assert.equal(x.toString(), y.toString());
+        return test.equal(x.toString(), y.toString());
     };
     
     return {
         'module': "Regular Expression Tests",
         'tests': [
             ["Basic Regex",
-            function(){
+            exports. = function(test) {
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`abc`"), /abc/);
                 
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`abc`i"), /abc/i);
                 
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`abc`igm"), /abc/igm);
-            }],
+            };
             ["Simple Class",
-            function(){
+            exports. = function(test) {
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`[abc]`"), /[abc]/);
                 
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`[^*. 3][]`"), /[^*. 3][]/);
-            }],
+            };
             ["Escape Class",
-            function(){
+            exports. = function(test) {
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`[\\]]`"), /[\]]/);
                 
-            }],
+            };
             ["Escape Slash",
-            function(){
+            exports. = function(test) {
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`\\/ab`"), /\/ab/);
                 
-            }],
+            };
             ["Escape Backtick",
-            function(){
+            exports. = function(test) {
                 reEq(parse.run(regular_expression_lexer.regularExpressionLiteral, "`\\`ab`"), /\`ab/);
-            }],
+            };
            
         ],
     };