+++ dist_node/parse/pattern_parser.js
@@ -78,7 +78,7 @@ var subPatternElements = eager(sepBy1(sep, either(topLevelPattern, subPattern)))
 (subPattern = Parser("Sub Pattern", attempt(nodea(enumeration(identifierPattern, between(punctuator("("), punctuator(
     ")"), subPatternElements)), ast_pattern.ArgumentsPattern.create))));
 var argumentElements = eager(sepBy(sep, topLevelPattern)),
-    selfPattern = next(punctuator("="), identifierPattern);
+    selfPattern = next(punctuator("="), choice(arrayPattern, objectPattern, asPattern, identifierPattern));
 (argumentList = Parser("Argument List", nodea(enumeration(argumentElements, optional(null, selfPattern)), (function(loc,
     elements, self) {
     return ast_pattern.ArgumentsPattern.create(loc, null, elements, self);