+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.1 - Feb 3, 2014
 * Removed use of custom ParserState for parsing
+* Removed support for unicode `\\` escaped sequences in identifiers which was
+  always lexed to the `\\` puctuator anyways.
 
 ## 0.0.0 - Feb 3, 2014
 * Fork from main Khepri codebase.