+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog #
 
+## 0.1.0 - Feb 19, 2014
+* Allow self unpack to be any of: array pattern, as pattern, object pattern,
+  identifier pattern.
+
 ## 0.0.1 - Feb 3, 2014
 * Removed use of custom ParserState for parsing
 * Removed support for unicode `\\` escaped sequences in identifiers which was